Refactored common module imports
