+++ src/cli.ts
@@ -4,7 +4,7 @@ import 'reflect-metadata';
 import CLIApplication from './app/cli-application.js';
 import GenerateCommand from './cli-command/generate-command/generate-command.js';
 import { HelpCommand, VersionCommand } from './cli-command/index.js';
-import { appContainer } from './common/container/container.js';
+import { appContainer } from './container/container.js';
 import { Component } from './types/index.js';
 
 const cli = new CLIApplication();
