+++ src/main.ts
@@ -1,5 +1,5 @@
 import Application from './app/application.js';
-import { appContainer } from './common/container/container.js';
+import { appContainer } from './container/container.js';
 import { Component } from './types/index.js';
 
 const app = appContainer.get<Application>(Component.Application);
